Prevent invalid synopsis part warning when default value is "false"
Fixes a set of warnings that look like this:

```
Warning: The `wp hmci import buried-treasure` command has an invalid synopsis part: --resume
Warning: The `wp hmci import buried-treasure` command has an invalid synopsis part: --disable_intermediate_images
```

Previously we were checking whether the default value of a parameter was
empty in order to define that arg as optional. This broke on args where
the default was `false`, because an argument can't have a default value
and be declared optional.