Remove "required" flag on iterator args
Since we've added synopsis generation in #80, the plugin will construct
an instance of each importer class in order to generate the synopsis for
them. Since this class instantiation doesn't have any args passed into
it, having the constructor die on missing required args breaks any
classes that depend on these. I believe this is the only instance of an
arg marked "required" left in the plugin.